2023/07/31
Damage 와 HP Integer 형으로 변경, 일부 애니메이션 프레임 수 축소, 허공클릭시 0.0.0으로 이동하는 상황 제거(플레이어 hit박스 지정하려다 Moveable이 이상하게 작동되는 것을 알아냄),
프로젝트 세팅에 PlayerHitBox 추가, Player에 PlayerHitBox(BoxCollision)추가하고 Collision 설정 추가, Player 관련 코드 점검

+++ Source/Portfolio_cpp/private/Enemy.cpp
@@ -82,7 +82,7 @@ void AEnemy::SetupPlayerInputComponent(UInputComponent* PlayerInputComponent)
 
 }
 
-void AEnemy::OnDamageProcess(float damage)
+void AEnemy::OnDamageProcess(int damage)
 {
 	//isDamaged = true;
 
@@ -96,9 +96,10 @@ void AEnemy::OnDamageProcess(float damage)
 	//피격애니몽타주 재생 
 	// 
 	//체력 감소 
-	HP -= damage;
-	UE_LOG(LogTemp, Warning, TEXT("HP:%.2f"),HP );
-	if (HP <= 0.0f)
+	
+	HP-= damage;
+	UE_LOG(LogTemp, Warning, TEXT("HP:%d"),HP );
+	if (HP <= 0)
 	{
 		isDead = true;
 		GetCapsuleComponent()->SetCollisionEnabled(ECollisionEnabled::NoCollision);
